split resolve.c, add a linker seam to facilitate tests
